chore: separate the db tables into separate files
