+++ src/libs/shared/src/db/schema/auth.ts
@@ -1,0 +1,32 @@
+import { boolean, pgTable, serial, text, timestamp, integer } from "drizzle-orm/pg-core";
+import type { InferInsertModel, InferSelectModel } from "drizzle-orm";
+
+// Users table
+export const users = pgTable("users", {
+  id: serial("id").primaryKey(),
+  userName: text("user_name").notNull().unique(),
+  emailAddress: text("email_address").notNull().unique(),
+  emailVerified: boolean("email_verified").notNull().default(false),
+  hashedPassword: text("hashed_password").notNull(),
+  createdAt: timestamp("created_at", { mode: "date", withTimezone: true }).notNull().defaultNow(),
+  updatedAt: timestamp("updated_at", { mode: "date", withTimezone: true })
+    .notNull()
+    .defaultNow()
+    .$onUpdate(() => new Date()),
+});
+
+export type User = InferSelectModel<typeof users>;
+export type NewUser = InferInsertModel<typeof users>;
+export type SanitizedUser = Omit<User, "hashedPassword">;
+
+// User auth sessions table
+export const sessions = pgTable("sessions", {
+  id: text("id").primaryKey(),
+  userId: integer("user_id")
+    .notNull()
+    .references(() => users.id, { onDelete: "cascade" }),
+  expiresAt: timestamp("expires_at", { mode: "date", withTimezone: true }).notNull(),
+});
+
+export type Session = InferSelectModel<typeof sessions>;
+export type NewSession = InferInsertModel<typeof sessions>;
